# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocSetu Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  city?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: string;
  phone?: string;
  gender?: string;
  city?: string;
}

export interface LoginInput {
  /** Email address or phone number */
  identifier: string;
  password: string;
}

export interface WorkerProfile {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  skills: string[];
  location: string;
  rating: number;
  reviewCount: number;
  isAvailable: boolean;
  isVerified: boolean;
  completedJobs: number;
  /** @nullable */
  experience?: string | null;
  languages: string[];
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  hourlyRate?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface WorkerProfileInput {
  skills: string[];
  location: string;
  experience?: string;
  bio?: string;
  hourlyRate?: number;
  languages?: string[];
}

export interface WorkerProfileUpdate {
  skills?: string[];
  location?: string;
  experience?: string;
  bio?: string;
  hourlyRate?: number;
  isAvailable?: boolean;
  languages?: string[];
}

export interface Skill {
  id: string;
  name: string;
  icon: string;
}

export interface Job {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  skill: string;
  location: string;
  status: string;
  customerId: number;
  /** @nullable */
  workerId?: number | null;
  /** @nullable */
  budget?: number | null;
  createdAt?: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface JobInput {
  title: string;
  description?: string;
  skill: string;
  location: string;
  budget?: number;
  workerId?: number;
}

export interface JobUpdate {
  status?: string;
  workerId?: number;
  budget?: number;
}

export interface Review {
  id: number;
  workerId: number;
  customerId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt?: string;
}

export interface ReviewInput {
  workerId: number;
  rating: number;
  comment?: string;
}

export interface SavedWorker {
  id: number;
  customerId: number;
  workerId: number;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  body: string;
  type: string;
  isRead: boolean;
  createdAt?: string;
}

export interface BuzzPost {
  id: number;
  userId: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  likesCount?: number;
  createdAt?: string;
  /** @nullable */
  userName?: string | null;
}

export interface BuzzInput {
  content: string;
  imageUrl?: string;
}

export interface EmergencyWorker {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  skills: string[];
  location: string;
  rating: number;
  reviewCount: number;
  isAvailable: boolean;
  isVerified?: boolean;
  completedJobs?: number;
  /** @nullable */
  phone?: string | null;
  estimatedArrival: number;
}

export interface EmergencyRequest {
  id: number;
  customerId: number;
  /** @nullable */
  workerId?: number | null;
  skill: string;
  location: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  estimatedArrival?: number | null;
  createdAt?: string;
}

export interface EmergencyRequestInput {
  skill: string;
  location: string;
  description?: string;
}

export interface ServicePricing {
  id: number;
  skill: string;
  serviceName: string;
  /** @nullable */
  serviceNameHi?: string | null;
  /** @nullable */
  serviceNameMr?: string | null;
  minPrice: number;
  avgPrice: number;
  maxPrice: number;
  unit: string;
}

export interface Course {
  id: number;
  title: string;
  category: string;
  description: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  duration: number;
  language: string;
  isFree: boolean;
  createdAt?: string;
}

export interface CourseProgress {
  id: number;
  userId: number;
  courseId: number;
  progressPercent: number;
  /** @nullable */
  completedAt?: string | null;
  certificateIssued?: boolean;
}

export interface CourseProgressInput {
  progressPercent: number;
}

export interface Availability {
  id: number;
  workerId: number;
  date: string;
  status: string;
  /** @nullable */
  note?: string | null;
}

export interface AvailabilityInput {
  date: string;
  status: string;
  note?: string;
}

export interface PortfolioItem {
  id: number;
  workerId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  mediaUrl: string;
  mediaType: string;
  /** @nullable */
  beforeUrl?: string | null;
  /** @nullable */
  afterUrl?: string | null;
  createdAt?: string;
}

export interface PortfolioItemInput {
  title: string;
  description?: string;
  mediaUrl: string;
  mediaType?: string;
  beforeUrl?: string;
  afterUrl?: string;
}

export interface SosEvent {
  id: number;
  userId: number;
  /** @nullable */
  latitude?: string | null;
  /** @nullable */
  longitude?: string | null;
  /** @nullable */
  locationAddress?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  /** @nullable */
  message?: string | null;
  createdAt?: string;
}

export interface SosInput {
  latitude?: string;
  longitude?: string;
  locationAddress?: string;
  emergencyContact?: string;
  message?: string;
}

export type ListWorkersParams = {
skill?: string;
location?: string;
available?: boolean;
gender?: string;
};

export type ListEmergencyWorkersParams = {
skill?: string;
location?: string;
};

export type ListPricingParams = {
skill?: string;
};

export type ListCoursesParams = {
category?: string;
language?: string;
};