# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocSetu Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /emergency/workers:
    get:
      operationId: listEmergencyWorkers
      tags: [emergency]
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available emergency workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyWorker"
  /emergency/request:
    post:
      operationId: createEmergencyRequest
      tags: [emergency]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyRequest"
  /emergency/requests:
    get:
      operationId: listMyEmergencyRequests
      tags: [emergency]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: My emergency requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyRequest"
...
  /availability/{workerId}:
    get:
      operationId: getWorkerAvailability
      tags: [availability]
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
  /availability:
    post:
      operationId: setAvailability
      tags: [availability]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        phone:
          type: string
        gender:
          type: string
        city:
          type: string
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Email address or phone number
        password:
          type: string
...
    Job:
      type: object
      required: [id, title, skill, location, status, customerId]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        skill:
          type: string
        location:
          type: string
        status:
          type: string
        customerId:
          type: integer
        workerId:
          type: ["integer", "null"]
        budget:
          type: ["number", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    JobInput:
      type: object
      required: [title, skill, location]
      properties:
        title:
          type: string
        description:
          type: string
        skill:
          type: string
        location:
          type: string
        budget:
          type: number
        workerId:
          type: integer
    JobUpdate:
      type: object
      properties:
        status:
          type: string
        workerId:
          type: integer
        budget:
          type: number
...
    EmergencyWorker:
      type: object
      required: [id, userId, name, skills, location, rating, reviewCount, isAvailable, estimatedArrival]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        isVerified:
          type: boolean
        completedJobs:
          type: integer
        phone:
          type: ["string", "null"]
        estimatedArrival:
          type: integer
    EmergencyRequest:
      type: object
      required: [id, customerId, skill, location, status]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        workerId:
          type: ["integer", "null"]
        skill:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        estimatedArrival:
          type: ["integer", "null"]
        createdAt:
          type: string
    EmergencyRequestInput:
      type: object
      required: [skill, location]
      properties:
        skill:
          type: string
        location:
          type: string
        description:
          type: string
...
    Availability:
      type: object
      required: [id, workerId, date, status]
      properties:
        id:
          type: integer
        workerId:
          type: integer
        date:
          type: string
        status:
          type: string
        note:
          type: ["string", "null"]
    AvailabilityInput:
      type: object
      required: [date, status]
      properties:
        date:
          type: string
        status:
          type: string
        note:
          type: string
...