# lib/api-zod/src/generated/types/emergencyWorker.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocSetu Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyWorker {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  skills: string[];
  location: string;
  rating: number;
  reviewCount: number;
  isAvailable: boolean;
  isVerified?: boolean;
  completedJobs?: number;
  /** @nullable */
  phone?: string | null;
  estimatedArrival: number;
}